fix: preserve native heatmap slider behavior
