+++ web/components/dashboard/visualization/adapters/echarts.ts
@@ -5,7 +5,6 @@ import { clearInteractionCommand, interactionCommandForRow } from '../interactio
 import { projectVisualizationHighlights } from '../highlight'
 import { baseOption } from './echarts/common'
 import { cartesianOption } from './echarts/cartesian'
-import { normalizeHeatmapRangeSelection } from './echarts/heatmap-range'
 import { hierarchyOption } from './echarts/hierarchy'
 import { polarOption } from './echarts/polar'
 import { pointOption } from './echarts/point'
@@ -100,7 +99,6 @@ class EChartsHandle implements RendererHandle {
   constructor(private readonly container: HTMLElement, private readonly frame: HTMLElement, private readonly chart: ECharts) {
     this.chart.on('click', this.handleClick)
     this.chart.on('brushSelected', this.handleBrushSelected)
-    this.chart.on('dataRangeSelected', this.handleDataRangeSelected)
   }
 
   mount(envelope: VisualizationEnvelope, context: RendererContext): void {
@@ -135,7 +133,6 @@ class EChartsHandle implements RendererHandle {
     this.readinessAbort = undefined
     this.chart.off('click', this.handleClick)
     this.chart.off('brushSelected', this.handleBrushSelected)
-    this.chart.off('dataRangeSelected', this.handleDataRangeSelected)
     this.chart.dispose()
     removeEChartsRendererFrame(this.container, this.frame)
   }
@@ -169,14 +166,6 @@ class EChartsHandle implements RendererHandle {
     }
   }
 
-  private readonly handleDataRangeSelected = (params: unknown) => {
-    const envelope = this.envelope
-    const selected = (params as { selected?: unknown }).selected
-    if (!envelope || !Array.isArray(selected)) return
-    const normalized = normalizeHeatmapRangeSelection(envelope, selected)
-    if (!normalized || normalized.every((value, index) => Object.is(value, selected[index]))) return
-    this.chart.dispatchAction({ type: 'selectDataRange', selected: normalized }, { silent: true })
-  }
 }
 
 export function brushSelectionCommands(envelope: VisualizationEnvelope, params: unknown) {
